Fix path after move
