+++ src/ng-http-loader.module.ts
@@ -8,35 +8,20 @@
  */
 
 import { NgModule } from '@angular/core';
-import { SpinnerComponent } from './spinner/spinner.component';
 import { CommonModule } from '@angular/common';
-import { PendingInterceptorService, PendingInterceptorServiceFactoryProvider } from './pending-interceptor.service';
-import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
-import { SPINKIT_COMPONENTS } from './spinkits';
-
-const PendingInterceptorServiceExistingProvider = {
-    provide: HTTP_INTERCEPTORS,
-    useExisting: PendingInterceptorService,
-    multi: true
-};
+import { NgHttpLoaderComponentsModule } from './components/ng-http-loader-components.module';
+import { NgHttpLoaderServicesModule } from './services/ng-http-loader-services.module';
 
 @NgModule({
-    declarations: [
-        SpinnerComponent,
-        SPINKIT_COMPONENTS,
-    ],
     imports: [
         CommonModule,
-        HttpClientModule,
+        NgHttpLoaderComponentsModule,
+        NgHttpLoaderServicesModule,
     ],
     exports: [
-        SpinnerComponent,
-        SPINKIT_COMPONENTS,
+        NgHttpLoaderComponentsModule,
+        NgHttpLoaderServicesModule,
     ],
-    providers: [
-        PendingInterceptorServiceExistingProvider,
-        PendingInterceptorServiceFactoryProvider,
-    ]
 })
 export class NgHttpLoaderModule {
 }
