+++ test/components/sk-double-bounce/sk-double-bounce.component.spec.ts
@@ -8,7 +8,7 @@
  */
 
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { SkDoubleBounceComponent } from '../../src/sk-double-bounce/sk-double-bounce.component';
+import { SkDoubleBounceComponent } from '../../../src/components/sk-double-bounce/sk-double-bounce.component';
 
 
 describe('SkDoubleBounceComponent', () => {
