+++ test/components/sk-spinner-pulse/sk-spinner-pulse.component.spec.ts
@@ -8,7 +8,7 @@
  */
 
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { SkSpinnerPulseComponent } from '../../src/sk-spinner-pulse/sk-spinner-pulse.component';
+import { SkSpinnerPulseComponent } from '../../../src/components/sk-spinner-pulse/sk-spinner-pulse.component';
 
 
 describe('SkSpinnerPulseComponent', () => {
