+++ test/components/sk-wave/sk-wave.component.spec.ts
@@ -8,7 +8,7 @@
  */
 
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { SkWaveComponent } from '../../src/sk-wave/sk-wave.component';
+import { SkWaveComponent } from '../../../src/components/sk-wave/sk-wave.component';
 
 
 describe('SkWaveComponent', () => {
